+++ src/compas_fab/backends/ros/messages/object_recognition_msgs.py
@@ -4,8 +4,9 @@ from .std_msgs import ROSmsg
 
 
 class ObjectType(ROSmsg):
-    """https://docs.ros.org/kinetic/api/object_recognition_msgs/html/msg/ObjectType.html
+    """https://docs.ros.org/en/api/object_recognition_msgs/html/msg/ObjectType.html
     """
+    ROS_MSG_TYPE = 'object_recognition_msgs/ObjectType'
 
     def __init__(self, key='key', db='db'):
         self.key = key