+++ src/compas_fab/backends/ros/messages/octomap_msgs.py
@@ -10,6 +10,7 @@ from .geometry_msgs import Pose
 
 class Octomap(ROSmsg):
     """https://docs.ros.org/kinetic/api/octomap_msgs/html/msg/Octomap.html"""
+    ROS_MSG_TYPE = 'octomap_msgs/Octomap'
 
     def __init__(self, header=None, binary=False, id='', resolution=0., data=None):
         self.header = header or Header()  # Header
@@ -21,6 +22,7 @@ class Octomap(ROSmsg):
 
 class OctomapWithPose(ROSmsg):
     """https://docs.ros.org/kinetic/api/octomap_msgs/html/msg/OctomapWithPose.html"""
+    ROS_MSG_TYPE = 'octomap_msgs/OctomapWithPose'
 
     def __init__(self, header=None, origin=None, octomap=None):
         self.header = header or Header()  # Header